feat: Add initial app structure and dependencies
This commit introduces the initial structure of the Guildo app, including the `home` feature with its associated page, controller, and binding. It also adds the `restaurants` controller, along with its binding, state, and provider.

The commit includes theming support with `VolcanoColors` and `VolcanoTheme`, leveraging `FlexColorScheme` for dynamic light and dark themes.

Furthermore, it configures initial routing using `RouteConfig` and `GetPage` for navigation. The `main.dart` file is also refactored to use `ScreenUtilInit` and `GetMaterialApp`, with support for multiple locales.

Finally the commit adds `ApiClient`, `Apis` and `GuildoConstants` for API interactions and define Restaurant models. It also adds custom types and custom widget extensions.