feat: Add Supabase integration, localization, and initial home page setup
This commit integrates Supabase as the backend service provider by utilizing `SupabaseConnector` and making it available in the `HomeBinding`.
The `supabase_connector` is also updated to be a `GetxService` instead of a class.
It also sets up `flutter_localizations` for internationalization and defines `supportedLocales` with French and US English locales, in `main.dart` file.

Furthermore, it introduces the initial setup for the home page, implemented in `home_page.dart` with a basic `Scaffold`.

The `pubspec.yaml` file has been modified to include the `flutter_localizations` dependency, and `pubspec.lock` has been updated accordingly.

Finally the commit adds `DatabaseTables` to define database tables name and `models.dart` to export `restaurant.dart`. It also creates `services.dart` to export all the services, and update `HomeController` to be injected in `HomeBinding`.